requirements.txt
cache
cachetools
python-dotenv
faiss-cpu
chainlit
langchain-experimental
groq==0.10.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.35
langchain-groq==0.1.9
langchain-openai==0.1.10
langchain-text-splitters==0.2.1
langchain-unstructured
llama-index
llama-parse
sentence-transformers
unstructured[all-docs]